requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-script
requests
flask-cors
pyjwt
uwsgi
psycopg2
